feat(auth): restore NextAuth changes split from docs batch
