+++ types/next-auth.d.ts
@@ -1,0 +1,27 @@
+import type { DefaultSession, DefaultUser } from 'next-auth'
+import type { DefaultJWT } from 'next-auth/jwt'
+import type { UserRole } from '@prisma/client'
+
+declare module 'next-auth' {
+  interface Session {
+    user: DefaultSession['user'] & {
+      id: string
+      email: string
+      name: string | null
+      role: UserRole
+      emailVerified: Date | null
+    }
+  }
+
+  interface User extends DefaultUser {
+    role: UserRole
+    emailVerified: Date | null
+  }
+}
+
+declare module 'next-auth/jwt' {
+  interface JWT extends DefaultJWT {
+    id: string
+    role: UserRole
+  }
+}
